ai/text-to-music/diffrhythm/requirements.txt
accelerate==1.4.0
torchdiffeq==0.2.5
torchaudio==2.6.0
x-transformers==2.1.2
transformers==4.49.0
librosa==0.10.2.post1
pyarrow==19.0.1
pandas==2.2.3
pylance==0.23.2
ema-pytorch==0.7.7
prefigure==0.0.10
bitsandbytes==0.45.3
muq==0.1.0
mutagen==1.47.0
pyopenjtalk==0.4.0
pykakasi==2.3.0
jieba==0.42.1
cn2an==0.5.23
pypinyin==0.53.0
onnxruntime==1.20.1
Unidecode==1.3.8
phonemizer==3.3.0
LangSegment==0.3.5
streamlit==1.29.1